chore: rename app
