+++ .gitlab-ci.yml
@@ -1,7 +1,7 @@
 variables:
-    TUTOR_PLUGIN: dash
-    TUTOR_PYPI_PACKAGE: tutor-dash
-    GITHUB_REPO: overhangio/tutor-dash
+    TUTOR_PLUGIN: deck
+    TUTOR_PYPI_PACKAGE: tutor-deck
+    GITHUB_REPO: overhangio/tutor-deck
 
 include:
     - project: "community/tutor-ci"
