+++ .hatch_build.py
@@ -16,7 +16,7 @@ class MetaDataHook(MetadataHookInterface):
 def load_about() -> dict[str, str]:
     about: dict[str, str] = {}
     with open(
-        os.path.join(HERE, "tutordash", "__about__.py"), "rt", encoding="utf-8"
+        os.path.join(HERE, "tutordeck", "__about__.py"), "rt", encoding="utf-8"
     ) as f:
         exec(f.read(), about)  # pylint: disable=exec-used
     return about
