+++ pyproject.toml
@@ -2,7 +2,7 @@
 # https://hatch.pypa.io/latest/config/build/
 
 [project]
-name = "tutor-dash"
+name = "tutor-deck"
 license = { text = "AGPL-3.0-only" }
 authors = [
   {name = "Edly"},
@@ -47,13 +47,13 @@ dev = [
 ]
 
 [project.entry-points."tutor.plugin.v1"]
-dash = "tutordash.plugin"
+deck = "tutordeck.plugin"
 
 # https://packaging.python.org/en/latest/specifications/well-known-project-urls/#well-known-labels
 [project.urls]
-Homepage = "https://github.com/overhangio/tutor-dash"
-Code = "https://github.com/overhangio/tutor-dash"
-Issues = "https://github.com/overhangio/tutor-dash/issues"
+Homepage = "https://github.com/overhangio/tutor-deck"
+Code = "https://github.com/overhangio/tutor-deck"
+Issues = "https://github.com/overhangio/tutor-deck/issues"
 
 # hatch-specific configuration
 [tool.hatch.metadata.hooks.custom]
@@ -66,8 +66,8 @@ build-backend = "hatchling.build"
 [tool.hatch.build.targets.sdist]
 # Disable strict naming, otherwise twine is not able to detect name/version
 strict-naming = false
-include = [ "/tutordash"]
+include = [ "/tutordeck"]
 exclude = ["tests*"]
 
 [tool.hatch.build.targets.wheel]
-packages = ["tutordash"]
+packages = ["tutordeck"]
