+++ tutordeck/plugin.py
@@ -13,17 +13,17 @@ from .server import app
 
 hooks.Filters.CONFIG_DEFAULTS.add_items(
     [
-        ("DASH_VERSION", __version__),
+        ("DECK_VERSION", __version__),
     ]
 )
 
 
 @click.group()
-def dash() -> None:
+def deck() -> None:
     pass
 
 
-@dash.command(name="run")
+@deck.command(name="run")
 @click.option("--host", default="127.0.0.1", show_default=True)
 @click.option("-p", "--port", default=3274, type=int, show_default=True)
 @click.option(
@@ -31,11 +31,11 @@ def dash() -> None:
     help="Enable development mode, with auto-reload and debug templates.",
 )
 @click.pass_obj
-def dash_run(obj: Context, host: str, port: int, dev: bool) -> None:
+def deck_run(obj: Context, host: str, port: int, dev: bool) -> None:
     """
-    Run the dash server.
+    Run the deck server.
     """
     app.run(obj.root, host=host, port=port, debug=dev, use_reloader=dev)
 
 
-hooks.Filters.CLI_COMMANDS.add_item(dash)
+hooks.Filters.CLI_COMMANDS.add_item(deck)
