+++ tutordeck/server/static/css/.gitignore
