+++ tutordeck/server/templates/_plugin_installed_list.html
@@ -7,7 +7,7 @@
         <div class="author">By {{ plugin.author }}</div>
         <div class="description">{{ plugin.description|safe }}</div>
     </div>
-    <div class="warning" id="warning-cookie-{{plugin.name}}">
+    <div class="warning" hx-preserve="true" id="warning-cookie-{{plugin.name}}">
         <img src="{{ url_for('static', filename='/img/Featured icon.svg')}}" alt="" title="Run launch platform for changes to this plugin to take effect">
     </div>
