+++ tutordeck/server/templates/index.html
@@ -4,16 +4,16 @@
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <meta name="description" content="Tutor Dash">
+    <meta name="description" content="Tutor Deck">
 
-    <title>Tutor Dash</title>
+    <title>Tutor Deck</title>
 
     <!-- Favicon -->
     <link rel="icon" type="image/x-icon" href="{{ url_for('static', filename='/img/favicon.png') }}">
     <!-- CSS -->
     <!-- <link rel="stylesheet" href="css/normalize.css"> -->
     <!-- <link rel="stylesheet" href="css/styles.css"> -->
-    <link href="{{ url_for('static', filename='css/dash.css') }}" rel="stylesheet">
+    <link href="{{ url_for('static', filename='css/deck.css') }}" rel="stylesheet">
     <script src="{{url_for('static', filename='js/htmx.min.js')}}"></script>
     <script src="{{url_for('static', filename='js/sse.js')}}"></script>
 </head>
@@ -99,7 +99,7 @@
         </div>
     </div>
     
-    <script src="{{ url_for('static', filename='js/dash.js') }}"></script>
+    <script src="{{ url_for('static', filename='js/deck.js') }}"></script>
     {% if sidebar_active_tab %}
     <script>
         document.getElementById('{{ sidebar_active_tab }}').classList.toggle('sidebar-tab-selected');
