+++ tutordeck/server/templates/plugin_installed.html
